% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tusk.R
\name{great_circle_dist}
\alias{great_circle_dist}
\title{Great circle distance}
\usage{
great_circle_dist(x1, x2)
}
\arguments{
\item{x1}{coordinate pair 1, as list with $lat and $lon}

\item{x2}{coordinate pair 2, as list with $lat and $lon}
}
\value{
The distance on the globe in kilometers.
}
\description{
Calculate the distance between two points as great circle distance
on the globe
}
\details{
Uses information from
http://janmatuschek.de/LatitudeLongitudeBoundingCoordinates, where
dist = arccos(sin(lat1) * sin(lat2) + cos(lat1) * cos(lat2) *
cos(lon1 - lon2)) * R
}
\examples{
great_circle_dist(list(lon = 0.3, lat = 40.91), list(lon = 1.3,
lat = 41.02))
}
