% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nearest_points}
\alias{nearest_points}
\title{Get 4 or 16 nearest gridpoints}
\usage{
nearest_points(coords, netcdf, data_set = "ts321", npoints = 4)
}
\arguments{
\item{coords}{the coordinates to find the nearest gridpoints for
as list with $lon and $lat in decimal degrees}

\item{netcdf}{the ncdf object holding the relevant data}

\item{data_set}{the kind of data used (see ?supported_sets) for details}
}
\value{
a list holding four or sixteen lists, with each holding a
lat/lot
}
\description{
Get four or sixteen nearest gridpoints for a lon/lat pair.
}
\details{
The returned list of gridpoint coordinates can be used for
interpolation
}

