% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tusk.R
\name{supported_sets}
\alias{supported_sets}
\title{Supported data sets}
\usage{
supported_sets()
}
\value{
a vector holding the short names of supported data sets
}
\description{
Reported supported data sets
}
\details{
ts401: CRU TS 4.01
ts322: CRU TS 3.22
ts321: CRU TS 3.21
spei22: SPEIbase v.2.2
pdsidai2011: scPDSI from Dai 2011a and 2011b
eobs160: E-OBS data daily data version 16.0
eobs140: E-OBS data daily data version 14.0
puhg_pet: Princeton University Hydroclimatology Group 61-yr (1948-2008) Potential Evaporation Dataset
}
\examples{
supported_sets()
}
