% Generated by roxygen2 (4.0.1): do not edit by hand
\name{supported_sets}
\alias{supported_sets}
\title{Supported data sets}
\usage{
supported_sets()
}
\value{
a vector holding the short names of supported data sets
}
\description{
Reported supported data sets
}
\details{
ts321: CRU TS 3.21
spei22: SPEIbase v.2.2
pdsidai2011: scPDSI from Dai 2011a and 2011b
}
\examples{
supported_sets()
}

