\name{supported_sets}
\alias{supported_sets}
\title{Supported data sets}
\usage{
  supported_sets()
}
\value{
  a vector holding the short names of supported data sets
}
\description{
  Reported supported data sets
}
\details{
  ts320: CRU TS 3.20 spei22: SPEIbase v.2.2
}
\examples{
supported_sets()
}

